lab4 服务端编程/lab4/requirements.txt
alembic==1.10.4
argcomplete==2.0.0
blinker==1.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
cryptography==39.0.1
dnspython==2.2.1
email-validator==1.2.1
Flask==2.3.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.20.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
pip==23.0.1
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.21.0
setuptools==66.0.0
SQLAlchemy==1.4.39
tensorflow==2.12.0
traitlets==5.8.1
typing_extensions==4.5.0
userpath==1.8.0
Werkzeug==2.3.2
wheel==0.38.4
WTForms==2.3.3
zipp==3.15.0
